+++ app/javascript/components/taxes/Taxes.vue
@@ -6,7 +6,7 @@
                     <el-tabs v-model="activeTab" style="margin-bottom: 30px;">
                         <el-tab-pane v-for="tax in taxes" :key="tax.name"
                                      v-bind:label="tax.name" v-bind:name="tax.name">
-
+                            <taxes-detail v-bind:taxDetail="tax.detail"/>
                         </el-tab-pane>
                     </el-tabs>
                 </el-card>
@@ -17,12 +17,14 @@
 
 <script>
   import SafeBody from '../base/SafeBody.vue'
+  import TaxesDetail from './TaxesDetail.vue'
   import taxesConnector from '../../model/taxes_connector'
 
   export default {
     name: 'Taxes',
     components: {
-      SafeBody
+      SafeBody,
+      TaxesDetail
     },
     data () {
       return {