+++ features/step_definitions/pageObject/tax_page_object.rb
@@ -1,0 +1,13 @@
+
+class TaxPageObject < APageObject
+
+  def initialize page
+    super
+  end
+
+  def find_total_for tax_name
+    @page.find("#tab-#{tax_name}").click
+    @page.find("#total").text
+  end
+
+end